{"schema":"decodex/commit/1","summary":"Split docs API route handlers","authority":"manual"}
